+++ lib/Api/GeneralContractsApi.php
@@ -481,6 +481,107 @@ class GeneralContractsApi
     }
 
     /**
+     * Operation terminate
+     *
+     * POST General/Contracts/{generalContractId}/terminate
+     *
+     * @param string $general_contract_id Contract identifier (required)
+     * @throws ApiException on non-2xx response
+     * @return \Secuconnect\Client\Model\ResultBoolean
+     */
+    public function terminate($general_contract_id)
+    {
+        list($response) = $this->terminateWithHttpInfo($general_contract_id);
+        return $response;
+    }
+
+    /**
+     * Operation terminateWithHttpInfo
+     *
+     * POST General/Contracts/{generalContractId}/terminate
+     *
+     * @param string $general_contract_id Contract identifier (required)
+     * @throws ApiException on non-2xx response
+     * @return array of \Secuconnect\Client\Model\ResultBoolean, HTTP status code, HTTP response headers (array of strings)
+     */
+    public function terminateWithHttpInfo($general_contract_id)
+    {
+        // verify the required parameter 'general_contract_id' is set
+        if ($general_contract_id === null || (is_array($general_contract_id) && count($general_contract_id) === 0)) {
+            throw new \InvalidArgumentException(
+                'Missing the required parameter $general_contract_id when calling terminate'
+            );
+        }
+        // parse inputs
+        $resourcePath = "/General/Contracts/{generalContractId}/terminate";
+        $httpBody = '';
+        $queryParams = [];
+        $headerParams = [];
+        $formParams = [];
+        $_header_accept = $this->apiClient->selectHeaderAccept(['application/json']);
+        if (!is_null($_header_accept)) {
+            $headerParams['Accept'] = $_header_accept;
+        }
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType([]);
+
+        // path params
+        if ($general_contract_id !== null) {
+            $resourcePath = str_replace(
+                "{" . "generalContractId" . "}",
+                $this->apiClient->getSerializer()->toPathValue($general_contract_id),
+                $resourcePath
+            );
+        }
+
+        // for model (json/xml)
+        if (isset($_tempBody)) {
+            $httpBody = $_tempBody; // $_tempBody is the method argument, if present
+        } elseif (count($formParams) > 0) {
+            $httpBody = $formParams; // for HTTP post (form)
+        }
+        for ($retries = 0; ; $retries++) {
+
+            // this endpoint requires OAuth (access token)
+            if (strlen($this->apiClient->getConfig()->getAccessToken()) !== 0) {
+                $headerParams['Authorization'] = 'Bearer ' . $this->apiClient->getConfig()->getAccessToken();
+            }
+
+            // make the API Call
+            try {
+                list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
+                    $resourcePath,
+                    'POST',
+                    $queryParams,
+                    $httpBody,
+                    $headerParams,
+                    '\Secuconnect\Client\Model\ResultBoolean',
+                    '/General/Contracts/{generalContractId}/terminate'
+                );
+
+                return [$this->apiClient->getSerializer()->deserialize($response, '\Secuconnect\Client\Model\ResultBoolean', $httpHeader), $statusCode, $httpHeader];
+            } catch (ApiException $e) {
+                switch ($e->getCode()) {
+                    case 200:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\ResultBoolean', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                    case 401:
+                        if ($retries < 1) {
+                            Authenticator::reauthenticate();
+                            continue 2;
+                        }
+                    default:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\ProductExceptionPayload', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                }
+
+                throw $e;
+            }
+        }
+    }
+
+    /**
      * Operation updateBankAccount
      *
      * POST General/Contracts/{generalContractId}/updateBankAccount